Add "mark" statement to script language.
Once a Scheme, Grammar or Vocab were created there was no way to either delete them or create new ones with the same code name. This creates a problem when repeatedly testing a script whilst making minor changes each time.

When the script first runs a named mark statement it has no effect. When run a second time all Scheme, Grammar and Vocab's that were created since the previous mark are removed from memory. So by placing a mark statement at the top of a script file, the file can be repeatedly run.