For Regnal and Hybrid calendar variants, allow for stated extended fields.
Regnal an Hybrid calendars need to explicitly state which fields and extended fields they re using since not all the component calendars will be have the same number and/or position of fields. If the component calendar uses the same name for its fields then they are automatically matched, otherwise a "match" sub statement must be given in the definition.