Change the default ranking for hybrid scheme fields.
By default, the first field in the record has the highest ranking - meaning it must always be present. In a hybrid scheme this is the would be the scheme index but, unless the date is ambiguous, this is not needed. So change the default ranking to put it after the supplied field names.