Cal library script, rewrite the script parser to be consistent.
In most places, statement text arguments can be names or primary values. Where specific value types are required, they are checked and converted if possible and an error issued if not.

The unused 'match' operator has been removed.

The regnal scheme sub-statement 'fixed' replaces the ',' separator between field and value with the '=' character. The English Regnal scheme "er" is updated with this change.