Cal library, create left padded pseudo style string corrected.
When the pseudo style string is modified to show it is to be padded to the left, it does this by reducing the existing pseudo style length to at lest one less than the existing string.

We now check that the existing string is not already less than the padded length.