Cal library, add the element sub-statement to the grammar definition statement.
The element sub-statement is used to introduce a new element name that does not exist in the underlying record. The 'alias field' sub-statement is used to match this field to one of the standard record fields. When the element is output using a format string, the supplied code is run to evaluate the required text.

The test file test/tscripts/suite/statements/grammar/element001.hcs added to confirm output.