Cal library, allow anonymous grammar sub-statement in scheme definition.
If the grammar sub-statement in a scheme definition does not have a grammar code but does have a body, then create an anonymous grammar for that scheme.

Change test script test/tscripts/suite/cals/astro001.hcs to declare an anonymous grammar with optional fields in place of the scheme definition optional sub-statement.