Cal library, ensure all schemes have an attached grammar.
Create a default grammar for schemes that don't have either a pre-defined or an anonymous grammar.

The first scheme to use a predefined grammar will set the default formats. Thereafter when the grammar is attached to more schemes the defaults are checked (by comparing the "def" control string) to ensure compatibility.

Some of the schemes in the default script have been temporarily removed (see scripts/caldefscripts.f2c).