HistoryCalTest Program, update to version 0.9.0
Command line no long needs -t flag, it will be ignored if given.

The command line can contain a list of files and directories. If a file is found it will be run as a test script. If it is a directory, all the contained files and subdirectories are run.

If the switch -n is given the test files following will be run without the standard library and default calendar schemes. The switch -l (lowercase L) will put the standard library back for the following files.

Moved all current test scripts to test/tscripts/suite/stdlib (from test/tscripts/suite).