Cal library, grammar class holds and checks base field names.
On attaching a grammar to a scheme, a list of the schemes base field names is checked with those held with the grammar. If they don't match, an error is generated.

If the grammar doesn't have a base field name list yet, then the it will be created the first time it is attached to a scheme.