Cal library, require script grammar statement to state fields.
Add fields sub-statement to all non-inherited grammar statements in the default library.

On creating a grammar object, we set all the properties and add the formats before we call the Grammar::construct member. This is a "run once - and only once" function that sets the grammar m_ok flag to true.

This prepares for setting up all the format sub-statements once all properties are known.