Cal script, comparing record value types now works.
Missing (short record) fields are assumed to be unknown (?). Blank fields (*) (coming soon) are also treated as unknown.

Records can only be compared for equal or unequal. Greater-than etc. is treated as an error.

Any value type not catered for is now treated as an error. It use to always return true.